test: 대표 카테고리 장소 업데이트 및 삭제 기능에 대한 테스트 이동
- 카테고리 테스트로 이동
